Fix multiple click card bug
If the user clicks a card more than once it will register a match and break the game.

Add code to check if the card has been clicked already and prevent a match caused by double clicking the card.

Add resetBoard function for when users double click a card so cards can be clicked again on the next attempt to find a match.