Update start game modal difficulty buttons
Create new JS file to handle button event listeners to better organise the code.

Add functionality to difficulty setting buttons on the start game modal.

Save a custom value for the countdown variable to the browser when a button is clicked. This changes the countdown speed of the game to accomodate for different difficulty settings.