Update start game modal button
Add function that checks whether a difficulty setting has been chosen.

If a difficulty setting was chosen then users will be taken to the game upon clicking the 'Find Your Coffee' button.

If not then the page displays an alert box asking users to pick a difficulty setting before playing the game.