Add interface.css file to handle the game timer
Remove CSS code that handles the styling of the timer from memory-game.css.

Add the code taken from memory-game.css to interface.css.

Add source tag for interface.css in memory-game.html.

Reduce the amount of code that needs to be written by creating a CSS file that will be used in all games.

Prevent code from being repeated unnecessarily which will better organise the code and file structure of the project.